requirements.txt
aiogram
python-dotenv
pyqiwip2p
requests